chore: refactor code (#96451)
